Added a delete button to reviews. Working on functionality

+++ Controllers/ReviewsController.cs
@@ -80,7 +80,14 @@ namespace DadsDayApp.Controllers
         // In the sample URL above it is the `5`. The "{id} in the [HttpDelete("{id}")] is what tells dotnet
         // to grab the id from the URL. It is then made available to us as the `id` argument to the method.
         //
+        // DELETE: api/Reviews/5
+        //
+        // Deletes an individual Review with the requested id. The id is specified in the URL
+        // In the sample URL above it is the `5`. The "{id} in the [HttpDelete("{id}")] is what tells dotnet
+        // to grab the id from the URL. It is then made available to us as the `id` argument to the method.
+        //
         [HttpDelete("{id}")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<IActionResult> DeleteReview(int id)
         {
             // Find this review by looking for the specific id
@@ -91,14 +98,31 @@ namespace DadsDayApp.Controllers
                 return NotFound();
             }
 
+            if (review.UserId != GetCurrentUserId())
+            {
+                // Make a custom error response
+                var response = new
+                {
+                    status = 401,
+                    errors = new List<string>() { "Not Authorized" }
+                };
+
+                // Return our error with the custom response
+                return Unauthorized(response);
+            }
+
             // Tell the database we want to remove this record
             _context.Reviews.Remove(review);
 
             // Tell the database to perform the deletion
             await _context.SaveChangesAsync();
 
-            // Return a copy of the deleted data
-            return Ok(review);
+            // return NoContent to indicate the update was done. Alternatively you can use the
+            // following to send back a copy of the deleted data.
+            //
+            // return Ok(review)
+            //
+            return NoContent();
         }
 
         // Private helper method that looks up an existing review by the supplied id
